requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
PyYAML>=6.0.1
joblib>=1.3.0

# API
fastapi>=0.110.0
uvicorn>=0.27.0

# Modeling
xgboost>=2.0.0

# Explainability / plots
matplotlib>=3.7.0
shap>=0.45.0

# Testing
pytest>=8.0.0


